--- aspnetcore/security/authentication/scaffold-identity/6.0sample/WebAuth/Areas/Identity/Pages/_ViewStart.cshtml
@@ -1,3 +1,0 @@
-﻿@{
-    Layout = "/Pages/Shared/_Layout.cshtml";
-}
