--- aspnetcore/security/authentication/scaffold-identity/6.0sample/WebAuth/Pages/Privacy.cshtml
@@ -1,8 +1,0 @@
-﻿@page
-@model PrivacyModel
-@{
-    ViewData["Title"] = "Privacy Policy";
-}
-<h1>@ViewData["Title"]</h1>
-
-<p>Use this page to detail your site's privacy policy.</p>
