--- aspnetcore/security/authentication/scaffold-identity/6.0sample/WebAuth/Pages/_ViewImports.cshtml
@@ -1,5 +1,0 @@
-﻿@using Microsoft.AspNetCore.Identity
-@using WebAuth
-@using WebAuth.Data
-@namespace WebAuth.Pages
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
