--- aspnetcore/security/authentication/scaffold-identity/6.0sample/WebAuth/Pages/_ViewStart.cshtml
@@ -1,3 +1,0 @@
-﻿@{
-    Layout = "_Layout";
-}
