--- aspnetcore/security/authentication/scaffold-identity/6.0sample/WebAuth/appsettings.Development.json
@@ -1,9 +1,0 @@
-{
-  "DetailedErrors": true,
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  }
-}
