--- aspnetcore/security/authentication/scaffold-identity/6.0sample/WebAuth/wwwroot/css/site.css
@@ -1,18 +1,0 @@
-html {
-  font-size: 14px;
-}
-
-@media (min-width: 768px) {
-  html {
-    font-size: 16px;
-  }
-}
-
-html {
-  position: relative;
-  min-height: 100%;
-}
-
-body {
-  margin-bottom: 60px;
-}
